refactor(typing): move protocols to _typing.py and update imports across modules
