+++ numpyro/distributions/copula.py
@@ -8,11 +8,11 @@ from jaxtyping import Array, ArrayLike, PRNGKeyArray
 
 from jax import lax, numpy as jnp
 
+from numpyro._typing import ConstraintLike, DistributionLike
 import numpyro.distributions.constraints as constraints
 from numpyro.distributions.continuous import Beta, MultivariateNormal, Normal
 from numpyro.distributions.distribution import Distribution
 from numpyro.distributions.util import clamp_probs, lazy_property, validate_sample
-from numpyro.typing import ConstraintLike, DistributionLike
 from numpyro.util import is_prng_key
